Chapter 5, lesson 22: Global base components In this lesson we learned how and where to use the Vue.component('componentName', component) function to register global base components that we can use in all other components without explicitly importing them.
We saw an example of how calling Vue.component with the global component in a single component caused the global component to be available everywhere because it was invoked on page load but discussed how this was bad practice as it would lead to needlessly bloated files.

We also added a new piece to the PageThreadShow component to display who created the post and when(using the now global AppDate component).