Chapter 5, lesson 25: Categories, collections of forums
This was a self guided lesson of sorts to see if I could implement categories using what I have learned thus far. I got everything functioning fairly quickly but did not know that the expectation was to update the home page to use the categories.

Things I goofed up:
- I had a reference to forums on the category list page.
- I did not defined a necessary computed property on the category list item page.
- I did not make the category page the new home page.

Beyond that there was a little cleanup from what I did that was able to be done once I proceeded through the lesson.