Class: The Vue.js master class Chapter: 6 Lesson: 26 Title: Migrating to Vuex Link: https://vueschool.io/lessons/migrating-to-vuex
In this lesson we added vuex to our project and replaced all of our uses of sourceData references over to using vuex via this.$store.state. This lesson directed me to take the Vuex for everyone course which I completed separately from this one and came back to this lesson. After going through the other class there were no surprises here though the covering of find and replace was a good call to make this work much faster.