Class: The Vue.js master class Chapter: 6 Lesson: 27 Title: Creating posts using Vuex Link: https://vueschool.io/lessons/creating-posts-using-vuex
In this lesson we converted the addPost method from PageThreadShow into several mutations and an action in the vuex store. We broke it into three separate mutations and called those mutations within a single action function. We also moved the call for that action to the PostEditors save function and removed the event listener in PageThreadShow, cleaning things up a good bit.