Class: The Vue.js master class Chapter: 6 Lesson: 29 Title: Grabbing the authenticated user Link: https://vueschool.io/lessons/grabbing-the-authenticated-user
Prior to this lesson I read the Vue style guidelines page which had a lot of great details. In this lesson we implemented an authenticated user and referenced their name and avatar image in the navbar using a currently hardcoded value. We used mapGetters to implement the getter in the navbar component.