Class: The Vue.js master class Chapter: 6 Lesson: 31 Title: Utility to count object properties Link: https://vueschool.io/lessons/utility-to-count-object-properties
In this lesson we created a utility function to count object properties and replaced a bunch of duplicated code with it. We learned that the common folder names to use for these are either helpers or utils. We went with utils. Also it was highlighted that it is better to export an object with the functions vs just exporting the functions so that where used you can only require the specific util functions you need.