Class: The Vue.js master class Chapter: 6 Lesson: 33 Title: Update a user profile Link: https://vueschool.io/lessons/update-a-user-profile
In the previous lesson we learned about how javascript passes some things by reference and some by value. In this lesson it was highlighted why it was important to know that as it showed bugs that can happen when you don't properly clone an object and instead pass it around by reference with actions and mutations. We created a UserProfileCard and a UserProfileCardEditor and allow the logged in user to update their profile. We added an action and mutation to the store to update the user in the store. Until we properly cloned the user, everything about the user was updated in realtime with or without clicking the submit button.