Class: The Vue.js master class Chapter: 7 Lesson: 38 Title: Add the route Link: https://vueschool.io/lessons/add-the-route
In this lesson we created the thread create route and wired it up to the "Start thread" button. I also fixed a bug where i had a typo in my createThread action. We learned that the ordering of routes defined in the router is very important as it tries to load the first matching route and if you use a dynamic property before an explicit property that the explicit property will be treated as a dynamic value. Specifically in this example trying to load a thread with an id of create. I left a comment in to highlight this so i remember.