Class: The Vue.js master class Chapter: 7 Lesson: 39 Title: Redirect on publish Link: https://vueschool.io/lessons/redirect-on-publish
In this lesson we wired up the publish button to redirect you to the newly published thread. We had to wrap our createThread function in a promise to return the newly created thread in order to be able to pass its id for the redirect to work. I am going to take a look at converting these over to async/await separate from this lesson. We also wired up the cancel button to take you to the forum.