Class: The Vue.js master class Chapter: 7 Lesson: 42 Title: Edit threads Link: https://vueschool.io/lessons/edit-threads
In this lesson we created a thread edit page and wired everything up to use the ThreadEdit component, making it a bit more dynamic. We also added a new route for editing threads, at this point viewing a thread and appending /edit at the end of the url will allow you to edit the thread.