Class: The Vue.js master class Chapter: 8 Lesson: 44 Title: Editing posts Link: https://vueschool.io/lessons/editing-posts
In this lesson we enabled the ability to edit posts, setting it to false at the end so it is not available, in the next lesson we will enable a toggle of some sort. The persist function we created was interesting in the way it was demonstrated initially to handle the promise.