Class: The Vue.js master class Chapter: 9 Lesson: 50 Title: Creating dynamic vuex getters using higher order functions Link: https://vueschool.io/lessons/creating-dynamic-vuex-getters-using-higher-order-functions
In this lesson we created some dynamic getters using currying and cleaned up the code a good bit with the two that we created. There was a lesson prior to this that went over higher order functions that did not involve any changes to the code but was very informative.