Class: The Vue.js master class Chapter: 10 Lesson: 62 Title: Fetch data for all pages Link: https://vueschool.io/lessons/fetch-data-for-all-pages
In this lesson we created fetch functions for single and multiple records for all of the types of resources we have. We implemented the necessary calls in create hooks in each page that requires the data to populate the data it needs, using if statements in the component to only render if the data is present.