Class: The Vue.js master class Chapter: 10 Lesson: 67 Title: Store New Threads in Firebase with Vuex Link: https://vueschool.io/lessons/store-new-threads-in-firebase-with-vuex
In this lesson we started persisting new threads into firebase in addition to posts. This is kind of messy for now as there is some duplicated code but it works. I also found and fixed a bug from an earlier lesson where i was referencing this.id instead of this.forumId on PageThreadCreate that was causing an error to show up in my console.