Class: The Vue.js master class Chapter: 11 Lesson: 72 Title: Share Logic Between Components with Mixins Link: https://vueschool.io/lessons/share-logic-between-components-with-mixins
In this lesson we added a mixins directory with a mixin to handle async data loaded functionality and added that functionality to the page components that load data asynchronously. Making sure everything is displayed at the same time. In this lesson it was highlighted that the naming for properties in mixins should reflect the mixin that they are defined in using a camelCase prefix followed by an underscore and the property name.