Class: The Vue.js master class Chapter: 11 Lesson: 73 Title: Display a loading indicator while components fetch async data Link: https://vueschool.io/lessons/display-a-loading-indicator-while-components-fetch-async-data
In this lesson we added a v-show statement to the app.vue component to have a single place to determine whether to display a page or not instead of repeating the same code in every component. We learned the difference between v-show and v-if(whether the component is rendered and hidden or not rendered at all). We also added a call in the created hook of the PageNotFound component so that it would properly render and know that it was ready as it was not using the shared mixin function that we used to set showPage = true. Additionally we used the beforeEach $router function to set showPage to false between each page navigation so that we could properly reset the showPage value to false.