Class: The Vue.js master class Chapter: 11 Lesson: 75 Title: Show a Progress Bar while Components Fetch Asynchronous Data Link: https://vueschool.io/lessons/show-a-progress-bar-while-complents-fetch-asynchronous-data
In this lesson we added a progress bar(at the very top of the page) when components are loading using the nprogress library. We used it in App.vue in our create hook before and after our router.beforeEach call so that it loads on initial page load and during navigation. We also extracted out a pageShow function to set showPage to true and finished our progress bar in that function.