Class: The Vue.js master class Chapter: 13 Lesson: 82 Title: Implement firebase authentication Link: https://vueschool.io/lessons/implement-firebase-authentication
In this lesson we wired our registration page up to firebase authentication using the registerWithEmailAndPassword function to create the user in firebase. The contract to this call has changed since the video was created so there was a bit of debugging to find where the users uid is in the response from firebase but other than that this was pretty straightforward.